requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
pytz>=2023.3
google-generativeai>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0

